Q&A Chatbot/requirements.txt
streamlit==1.38.0
langchain==0.2.12
langchain-core==0.2.28
langchain-community==0.2.6
langchain-huggingface==0.0.3
transformers==4.43.3
huggingface_hub==0.24.6
pydantic==2.7.4
python-dotenv
